feat: create new toptokens and tokenPrice query and function
